Fix unsafe lints
