+++ .rubocop.yml
@@ -116,6 +116,9 @@ Style/Documentation:
 Style/EmptyClassDefinition:
   Exclude:
     - 'lib/suma/*/model.rb' # Causes issues with how Sequel establishes connections
+Style/FileOpen:
+  Exclude:
+    - 'spec/**/*' # No need to worry about this in specs
 Style/FormatString:
   EnforcedStyle: percent
 Style/NumericPredicate:
