+++ lib/sequel/plugins/immutable.rb
@@ -2,10 +2,11 @@
 
 # Plugin to make an object immutable after creation.
 module Sequel::Plugins::Immutable
-end
+  __ = 0
 
-module Sequel::Plugins::Immutable::InstanceMethods
-  def before_update
-    raise FrozenError, "cannot modify immutable model"
+  module InstanceMethods
+    def before_update
+      raise FrozenError, "cannot modify immutable model"
+    end
   end
 end
