+++ lib/suma/liquid.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "liquid"
+
+module Suma::Liquid
+end
