+++ lib/suma/messages.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+module Suma::Messages
+end
