+++ lib/suma/service/auth.rb
@@ -2,16 +2,18 @@
 
 require "suma/yosoy"
 
-module Suma::Service::Auth; end
+module Suma::Service::Auth
+  __ = 0
 
-class Suma::Service::Auth::Middleware < Suma::Yosoy::Middleware
-  def serialize_into_session(session, _env) = session.token
+  class Middleware < Suma::Yosoy::Middleware
+    def serialize_into_session(session, _env) = session.token
 
-  def serialize_from_session(key, _env)
-    Suma::Member::Session.dataset.valid[token: key]
-  end
+    def serialize_from_session(key, _env)
+      Suma::Member::Session.dataset.valid[token: key]
+    end
 
-  def inactivity_timeout
-    Suma::Service.max_session_age
+    def inactivity_timeout
+      Suma::Service.max_session_age
+    end
   end
 end
